REP-62: update test configuration
